requirements.txt
Django==3.2
djangorestframework==3.12.2
gunicorn==20.0.4
web3==4.9.2
psycopg2-binary==2.8.4
swaggerpy==0.2.1
pip==19.3.1
bcrypt==3.1.7
python-gettext==4.0 
